refactor: Separate TriggerV1List, lazyload it
