+++ src/lib/SideBars/Scripts/TriggerV1Data.svelte
